+++ components/index.tsx
@@ -124,4 +124,6 @@ export { default as BaseData } from './base-data'
 
 export { default as Link } from './link'
 
+export { default as TreeSelect } from './tree-select'
+
 export { default as ColorPicker } from './color-picker'